worker/requirements.txt
paho-mqtt==2.1.0
mysql-connector-python==9.1.0